+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### 2.3.5 (Unreleased)
 
 - [#542](https://github.com/influxdata/clockface/pull/542): Add `tabIndex` and `onKeyDown` props to `SelectableCard` to enable keyboard interactivity
+- [#541](https://github.com/influxdata/clockface/pull/541): Improve `testID` usability in `TreeNav` component
 
 #### 2.3.4 (2020-08-17)